requirements.txt
alembic==1.4.3
atomicwrites==1.4.0
attrs==20.3.0
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
colorama==0.4.4
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
# mysqlclient==2.0.1
pymysql==1.0.2[
numpy==1.19.4
opencv-python==4.4.0.46
packaging==20.7
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
python-editor==1.0.4
redis==3.5.3
rq==1.6.1
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.2
waitress==1.4.4
Werkzeug==1.0.1
WTForms==2.3.3
